Player Controller Rework Begun
I done started remaking the character controller. It isn't too different, but I'm using the CharacterController component instead of a rigidbody. This takes care of some problems like walking up slopes, and it isn't too hard to implement features that came naturally with rigidbodies.
There's still more work to do before I'm satisfied with the controller, and it also needs some fine tuning so that the physics feel good.