Completed Player Controller Rework
The all new character controller is here, complete with a new prefab! I didn't alter the previous controller, so if for some reason the one I've made just doesn't work, then the old one can be used a bit.
I'm not sure if I got the Physics quite right, but it would only take a bit of tweaking to alter them.