Updated Player Controller
You thought I was done with this thing?
NONSENSE!!!

In this exciting installment of "Zack Screws With A Player Controller Because That Is All He Knows How To Do," I've accomplished such feats as:
- Putting the movement stuff in the update function.
-Made sliding down slopes a thing.
-Threw away unnecessary code.
-Sold off the compass that was taking up space.
-Gave the variables that I added some makeup.
-Got rid of the rigidbody that someone put on my child (I think it was Rob but I'm not sure, could've been me).

Up next I'll be messing with lore spots and dialogue. I'd also like to take a look at the orb holding mechanics if I have enough spare time. I don't know for sure, we'll see what happens.