Basic Dialogue System Implemented
With this, you can put dialogue on objects. The dialogue comes with a name, and by adding a "look point" the camera will turn to face where it should. In addition, the text fades in as the object "speaks."
Branching dialogue, skippable typing, and conversations between NPCs will be implemented at a later date.